backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
Flask-JWT-Extended==5.0.1
Werkzeug==3.1.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
mysql-connector-python==9.0.0
